detection-server/requirements.txt
ultralytics
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
pillow
nats-py

# PyTorch (CUDA is autodetected from the index)
--extra-index-url https://download.pytorch.org/whl/cu130
torch
torchvision
torchaudio
